specs: fixed TempFile usage for file fixtures
In some randomized seeds a temp file was being removed and no content
was being uploaded, since the file didn't exist, and the file content
wasn't being displayed on the screen as expected.

Since there's no way of controlling when GC is going to remove the
temp files, the only way is trying to keep the reference on hold to
be usable. The solution for this was to simply return the TempFile
instance from `to_file_fixture` instead of the path or base name.

Signed-off-by: Vítor Avelino <contact@vitoravelino.me>