js: added eslint
ESLint rules based on Portus have been added and most of the issues were
fixed. The dashboard.js file was the only one postponed because there
are pending PRs touching it and since there are a lot of changes, it
would cause a lot of conflicts.

Also moved ESLint, rubocop and brakeman to run before specs to fail
faster if there's a styling/security issue.

Signed-off-by: Vítor Avelino <vavelino@suse.com>