Make LDAP/External Auth Admin group name configurable
The internal OpenLDAP server utilizes a group named "Administrators" to give the admin user of Velum admin permissions to the cluster via a clusterrolebinding.

As a result, when using external authentication, any user in a Administrators group in an external auth server will additionally gain admin access. That is not intentional.

This change adds the ability configure the group name used within the settings section of Velum, so that the name used can be set to one that does not conflict with group names within existing corporate identity systems.