Simplify external cert fixture filenames
Moved external cert testing related fixture files into their own directory and simpified their filenames.

Dropped the pem extension and moved to a more clear key and crt extension.

Shortened both the documentation on the testing variables and the names of the variables themselves so that it is easier to see what is going on.