Get certchain checking working
Updated the external cert test certs generation script to copy over the second intermediate CA. Regenerated all the external_certs as a result of this so that it can be included.

Added calls to the SystemCertificate model to set system certificates in order to do testing of the chain checking in rspec. Attempted to use capybara instead of this but couldn't get it to work.

SHA1 digest is failing to test currently in this commit. It looks to be some sort of glitch with capybara as it tests manually without any issue.